daily/20200205/example_gcp_pubsub/requirements.lock.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
gcloud==0.18.3
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-pubsub==1.1.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.0
httplib2==0.17.0
idna==2.8
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
